Refactor + handle bck button visibility
